+++ test/setup.attest.ts
@@ -1,0 +1,6 @@
+import { setup, teardown } from "@ark/attest"
+
+export default (): (() => void) => {
+  setup({})
+  return teardown
+}