+++ vitest.config.ts
@@ -3,6 +3,7 @@ import { defineConfig } from "vitest/config"
 export default defineConfig({
   test: {
     include: ["test/**/*.test.ts"],
+    globalSetup: ["test/setup.attest.ts"],
     typecheck: {
       enabled: true,
       include: ["test/**/*.test-d.ts"],